+++ core/services/gateway/handlers/vault/aggregator_test_helpers.go
@@ -31,7 +31,6 @@ func makeSignedCreateSecretsResponse(t *testing.T, requestID string, numSigners 
 	t.Helper()
 
 	createResp := &vaultcommon.CreateSecretsResponse{
-		RequestId: requestID,
 		Responses: []*vaultcommon.CreateSecretResponse{
 			{
 				Id: &vaultcommon.SecretIdentifier{
@@ -44,7 +43,7 @@ func makeSignedCreateSecretsResponse(t *testing.T, requestID string, numSigners 
 			},
 		},
 	}
-	payload, err := vaultutils.ToCanonicalJSON(createResp)
+	payload, err := vaultutils.ToCanonicalJSON(createResp, false)
 	require.NoError(t, err)
 
 	return makeSignedVaultResponse(t, vaulttypes.MethodSecretsCreate, requestID, json.RawMessage(payload), numSigners)