+++ core/services/workflows/cmd/cre/examples/compile_test.go
@@ -1,5 +1,11 @@
 package examples
 
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/legacy/data_feeds
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/v2/http_read
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/v2/simple_cron
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/v2/simple_cron_with_config
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/v2/simple_cron_with_secrets
+//go:generate go run ../../../../../internal/testutils/wasmtest/generator/main.go -pkg core/services/workflows/cmd/cre/examples/v2/empty
 import (
 	"path/filepath"
 	"testing"
@@ -12,6 +18,7 @@ import (
 const pathPrefix = "core/services/workflows/cmd/cre/examples"
 
 func Test_AllExampleWorkflowsCompileToWASM(t *testing.T) {
+	t.Parallel()
 	paths := []string{
 		"legacy/data_feeds",
 		"v2/http_read",
@@ -23,7 +30,7 @@ func Test_AllExampleWorkflowsCompileToWASM(t *testing.T) {
 	for _, path := range paths {
 		t.Run(path, func(t *testing.T) {
 			t.Parallel()
-			binary := wasmtest.CreateTestBinary(t, filepath.Join(pathPrefix, path), false)
+			binary := wasmtest.GetTestBinary(t, filepath.Join(pathPrefix, path), false)
 			require.NotEmpty(t, binary)
 		})
 	}