+++ system-tests/tests/test-helpers/t_helpers.go
@@ -358,6 +358,7 @@ type WorkflowRegistrationConfig struct {
 	WorkflowRegistryVersion *semver.Version
 	ChainID                 uint64
 	DonID                   uint64
+	DonFamily               string
 	ContainerTargetDir      string
 	SethClient              *seth.Client
 	Attributes              []byte
@@ -637,6 +638,7 @@ func registerWorkflowErr(ctx context.Context, wfRegCfg *WorkflowRegistrationConf
 		wfRegCfg.WorkflowRegistryAddr,
 		wfRegCfg.WorkflowRegistryVersion,
 		wfRegCfg.DonID,
+		wfRegCfg.DonFamily,
 		wfRegCfg.WorkflowName,
 		binaryURL,
 		configURL,
@@ -730,6 +732,7 @@ func CompileAndDeployWorkflow[T WorkflowConfig](t *testing.T,
 		WorkflowRegistryVersion: workflowRegistryAddress.Version,
 		ChainID:                 registryChainSelector,
 		DonID:                   testEnv.Dons.MustWorkflowDON().ID,
+		DonFamily:               testEnv.Dons.MustWorkflowDON().DonFamily,
 		ContainerTargetDir:      creworkflow.DefaultWorkflowTargetDir,
 		SethClient:              testEnv.CreEnvironment.Blockchains[0].(*evm.Blockchain).SethClient,
 		Attributes:              cfg.attributes,